Allow Surfaces larger than 2GiB
Previously creating surfaces larger than 2^31 bytes would cause an int32_t overflow on data size calculation, throwing a bad_alloc exception.

This commit converts mRowBytes, mInc and mRowInc to size_t, which causes size calculations to be performed in size_t before the result is used for array allocation and pointer arithmetic. Also, sites where these variables are used have been converted to avoid loss of precision.

+++ include/cinder/msw/CinderMswGdiPlus.h
@@ -52,7 +52,7 @@ namespace cinder {
 
 class SurfaceConstraintsGdiPlus : public SurfaceConstraints {
 	virtual SurfaceChannelOrder getChannelOrder( bool alpha ) const { return ( alpha ) ? SurfaceChannelOrder::BGRA : SurfaceChannelOrder::BGR; }
-	virtual int32_t				getRowBytes( int requestedWidth, const SurfaceChannelOrder &sco, int elementSize ) const { int32_t result = requestedWidth * elementSize * sco.getPixelInc(); result += ( result & 3 ) ? ( 4 - (result&3) ) : 0; return result; }
+	virtual size_t				getRowBytes( int requestedWidth, const SurfaceChannelOrder &sco, int elementSize ) const { size_t result = requestedWidth * elementSize * sco.getPixelInc(); result += ( result & 3 ) ? ( 4 - (result&3) ) : 0; return result; }
 };
 
 } // namespace cinder
