+++ src/cinder/cocoa/CinderCocoa.mm
@@ -667,7 +667,7 @@ SurfaceChannelOrder SurfaceConstraintsCgBitmapContext::getChannelOrder( bool alp
 	return ( alpha ) ? SurfaceChannelOrder::RGBA : SurfaceChannelOrder::RGBX;
 }
 
-int32_t SurfaceConstraintsCgBitmapContext::getRowBytes( int requestedWidth, const SurfaceChannelOrder &sco, int elementSize ) const
+size_t SurfaceConstraintsCgBitmapContext::getRowBytes( int requestedWidth, const SurfaceChannelOrder &sco, int elementSize ) const
 {
 	return requestedWidth * elementSize * 4;
 }
