+++ src/cinder/ip/Threshold.cpp
@@ -33,8 +33,8 @@ template<typename T>
 void thresholdImpl( SurfaceT<T> *surface, T value, const Area &area )
 {
 	const Area clippedArea = area.getClipBy( surface->getBounds() );
-	int32_t rowBytes = surface->getRowBytes();
-	uint8_t pixelInc = surface->getPixelInc();
+	size_t rowBytes = surface->getRowBytes();
+	size_t pixelInc = surface->getPixelInc();
 	uint8_t redOffset = surface->getRedOffset(), greenOffset = surface->getGreenOffset(), blueOffset = surface->getBlueOffset();
 	T maxValue = CHANTRAIT<T>::max();
 	for( int32_t y = clippedArea.getY1(); y < clippedArea.getY2(); ++y ) {
@@ -55,11 +55,11 @@ void thresholdImpl( const SurfaceT<T> &srcSurface, T value, const Area &srcArea,
 	const Area &area( srcDst.first );
 	const ivec2 &dstOffset( srcDst.second );
 
-	int32_t srcRowBytes = srcSurface.getRowBytes();
-	int8_t srcPixelInc = srcSurface.getPixelInc();
+	size_t srcRowBytes = srcSurface.getRowBytes();
+	size_t srcPixelInc = srcSurface.getPixelInc();
 	uint8_t srcRedOffset = srcSurface.getRedOffset(), srcGreenOffset = srcSurface.getGreenOffset(), srcBlueOffset = srcSurface.getBlueOffset();
-	int32_t dstRowBytes = dstSurface->getRowBytes();
-	int8_t dstPixelInc = dstSurface->getPixelInc();
+	size_t dstRowBytes = dstSurface->getRowBytes();
+	size_t dstPixelInc = dstSurface->getPixelInc();
 	uint8_t dstRedOffset = dstSurface->getRedOffset(), dstGreenOffset = dstSurface->getGreenOffset(), dstBlueOffset = dstSurface->getBlueOffset();
 	const T maxValue = CHANTRAIT<T>::max();
 	for( int32_t y = 0; y < area.getHeight(); ++y ) {
@@ -82,8 +82,8 @@ void thresholdImpl( const ChannelT<T> &srcChannel, T value, const Area &srcArea,
 	const Area &area( srcDst.first );
 	const ivec2 &dstOffset( srcDst.second );
 
-	int8_t srcInc = srcChannel.getIncrement();
-	int8_t dstInc = dstChannel->getIncrement();
+	size_t srcInc = srcChannel.getIncrement();
+	size_t dstInc = dstChannel->getIncrement();
 	const T maxValue = CHANTRAIT<T>::max();
 	for( int32_t y = 0; y < area.getHeight(); ++y ) {
 		T *dstPtr = dstChannel->getData( ivec2( area.getX1(), y ) + dstOffset );
@@ -129,8 +129,8 @@ void calculateAdaptiveThreshold( const ChannelT<T> *srcChannel, typename CHANTRA
 	int32_t imageHeight = srcChannel->getHeight();
 
 	int s2 = windowSize / 2;
-	uint8_t srcInc = srcChannel->getIncrement();
-	uint8_t dstInc = dstChannel->getIncrement();
+	size_t srcInc = srcChannel->getIncrement();
+	size_t dstInc = dstChannel->getIncrement();
 
 	SUMT comparisonMult = static_cast<SUMT>( ( 1.0f - percentageDelta ) * 256 );
 	const T maxValue = CHANTRAIT<T>::max();
@@ -176,8 +176,8 @@ void calculateAdaptiveThresholdZero( const ChannelT<T> *srcChannel, typename CHA
 	int32_t imageWidth = srcChannel->getWidth();
 	int32_t imageHeight = srcChannel->getHeight();
 	int s2 = windowSize / 2;
-	uint8_t srcInc = srcChannel->getIncrement();
-	uint8_t dstInc = dstChannel->getIncrement();
+	size_t srcInc = srcChannel->getIncrement();
+	size_t dstInc = dstChannel->getIncrement();
 
 	// perform thresholding
 	for( int32_t j = 0; j < imageHeight; j++ ) {
@@ -220,8 +220,8 @@ template<typename T>
 void calculateIntegralImage( const ChannelT<T> &channel, typename CHANTRAIT<T>::Accum *integralImage )
 {
 	int32_t imageWidth = channel.getWidth(), imageHeight = channel.getHeight();
-	int32_t srcRowBytes = channel.getRowBytes();
-	uint8_t srcInc = channel.getIncrement();
+	size_t srcRowBytes = channel.getRowBytes();
+	size_t srcInc = channel.getIncrement();
 	const T *src = channel.getData();
 	/*for( int32_t i = 0; i < imageWidth; i++ ) {
 		// reset this column sum
